refactor(convention): 패키지 구조 변경
- web.controller.rest
