+++ src/main/java/com/woowacourse/zzinbros/common/web/controller/MainController.java
@@ -6,15 +6,9 @@ import com.woowacourse.zzinbros.user.service.UserService;
 import com.woowacourse.zzinbros.user.web.support.SessionInfo;
 import com.woowacourse.zzinbros.user.web.support.UserSession;
 import org.springframework.data.domain.Sort;
-import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-import java.time.Instant;
 
 import static org.springframework.data.domain.Sort.Direction;
 import static org.springframework.data.domain.Sort.by;
@@ -41,13 +35,5 @@ public class MainController {
     @GetMapping("/entrance")
     public String enter() {
         return "entrance";
-    }
-
-    @RequestMapping(path = "/datetime", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-    @ResponseBody
-    public String datetimeService() {
-        return "{ \"datetime\": \""
-                + Instant.now().toString()
-                + "\" }";
     }
 }
